% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{ancestor_aid}
\alias{ancestor_aid}
\title{Ancestor Adjustment Identification Distance between two DAG / CPDAG adjacency matrices (dense)}
\usage{
ancestor_aid(true_adjacency, guess_adjacency, edge_direction)
}
\description{
Ancestor Adjustment Identification Distance between two DAG / CPDAG adjacency matrices (dense)
}
