% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{children_of_first_node}
\alias{children_of_first_node}
\title{Gets the children of the first node, sanity check for row/column major loading}
\usage{
children_of_first_node(true_adjacency, edge_direction)
}
\description{
Gets the children of the first node, sanity check for row/column major loading
}
