% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{oset_aid}
\alias{oset_aid}
\title{Optimal Adjustment Intervention Distance between two DAG / CPDAG adjacency matrices (dense)}
\usage{
oset_aid(true_adjacency, guess_adjacency, edge_direction)
}
\description{
Optimal Adjustment Intervention Distance between two DAG / CPDAG adjacency matrices (dense)
}
