% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{parent_aid}
\alias{parent_aid}
\title{Parent Adjustment Intervention Distance between two DAG / CPDAG adjacency matrices (dense)}
\usage{
parent_aid(true_adjacency, guess_adjacency, edge_direction)
}
\description{
Parent Adjustment Intervention Distance between two DAG / CPDAG adjacency matrices (dense)
}
