% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{shd}
\alias{shd}
\title{Structural Hamming Distance between two DAG / CPDAG adjacency matrices (dense).
Does not take \code{edge_direction} argument, as SHD reads the matrix 1:1.}
\usage{
shd(true_adjacency, guess_adjacency)
}
\description{
Structural Hamming Distance between two DAG / CPDAG adjacency matrices (dense).
Does not take \code{edge_direction} argument, as SHD reads the matrix 1:1.
}
