% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{sid}
\alias{sid}
\title{Structural Intervention Distance between two DAG adjacency matrices (dense)}
\usage{
sid(true_adjacency, guess_adjacency, edge_direction)
}
\description{
Structural Intervention Distance between two DAG adjacency matrices (dense)
}
